made change to make work

+++ src/main/java/org/piki/App.java
@@ -13,7 +13,7 @@ public class App {
 
         long a = 1, b = 1;
         for (int i=2; i<n; i++) {
-            long t = a;
+            int t = a;
             a = b;
             b = b + t;
         }
